feat: Add 'static lifetime to RefreshTokenService and RefreshTokenRepository
This commit adds the 'static lifetime constraint to the RefreshTokenService and RefreshTokenRepository traits in the refresh_token.rs file. By adding the 'static lifetime, it ensures that the implementations of these traits can be used for the entire duration of the program. This change improves the flexibility and usability of the RefreshTokenService and RefreshTokenRepository traits.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>