feat: Add serial number parameter to create_refresh_token function
This commit adds a new parameter `serial_number` to the `create_refresh_token` function in both the `RefreshTokenService` and `RefreshTokenRepository` traits. This parameter allows for the creation of refresh tokens with a specific serial number. The addition of this parameter enhances the flexibility and customization of the token creation process.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>