feat: Add PostgresRefreshTokenRepository implementation
This commit adds the implementation of the `RefreshTokenRepository` trait for the PostgreSQL database in the `PostgresRefreshTokenRepository` struct. It includes the `create_refresh_token` method for creating refresh tokens in the database. This addition enhances the token infrastructure by providing support for storing and retrieving refresh tokens in a PostgreSQL database.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>