feat: Add `renew_tokens` method to ProviderTokenService
This commit adds a new method `renew_tokens` to the `ProviderTokenService` trait in the `provider_token_service.rs` file. The `renew_tokens` method takes a refresh token as input and returns a future that resolves to a result of `Tokens` or `CreateTokensError`. This method allows for the renewal of tokens using a refresh token.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>