Implement/provider token (#4)
* feat: Add Tokens struct for access and refresh tokens

The code changes add a new struct called `Tokens` which contains the `access_token` and `refresh_token` fields. This struct is used to represent a pair of tokens. Additionally, an enum `CreateTokensError` is introduced with a variant `ProviderNotFound` to handle the case when the provider is not found.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add provider_token_service module for handling provider-specific token authentication

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add `TokenProviderManager` module for managing token providers

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add Default implementation for TokenProviderManager

This commit adds a Default implementation for the TokenProviderManager struct, allowing it to be created with default values. This is useful when initializing the TokenProviderManager without any specific provider token service. The Default implementation calls the new() method to create a new instance of TokenProviderManager.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add error variants to CreateTokensError enum

This commit adds two new error variants, `ProviderError` and `InvalidToken`, to the `CreateTokensError` enum in the `Token` module. These variants handle cases where the provider returns an error or the token is invalid.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add cookie and httpmock dependencies

The code changes in the `Cargo.toml` file add the `cookie` and `httpmock` dependencies to the project. This is done to enable the use of cookies and to facilitate HTTP mocking in the codebase.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: create `BambuLab` provider

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: wip on domain

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Refactor HTTP routes for token creation and retrieval

The code changes in the `http.rs` file refactor the HTTP routes for creating and retrieving tokens. The routes are now organized using the `Router` struct from the `axum` crate. This improves code readability and maintainability.

```

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: add Clone trait

The code changes in the `bambulab_provider.rs` file introduce a new struct called `BambuLabProviderTokenService`. This struct is responsible for handling provider-specific token authentication. It includes an HTTP client and a `Clone` implementation. This refactor improves code organization and maintainability.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: implement token provider manager in application

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Add `renew_tokens` method to ProviderTokenService

This commit adds a new method `renew_tokens` to the `ProviderTokenService` trait in the `provider_token_service.rs` file. The `renew_tokens` method takes a refresh token as input and returns a future that resolves to a result of `Tokens` or `CreateTokensError`. This method allows for the renewal of tokens using a refresh token.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

* feat: Refactor HTTP routes for token creation and retrieval

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>

---------

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>