feat: Add AccessToken service and implementation
This commit adds the `AccessTokenService` trait and its implementation `AccessTokenServiceImpl` in the `token` module. The service provides a method `create_access_token` for creating access tokens. Additionally, a new module `services` is introduced to contain the service implementation. This enhancement improves the token domain by separating the service logic from the repository and models.

Signed-off-by: nathaelbonnal <pro.nathaelbonnal@gmail.com>